requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
attrs==24.2.0
cachetools==5.5.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dataclasses-json==0.5.14
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.99.1
fpdf==1.7.2
frozenlist==1.5.0
google-ai-generativelanguage==0.4.0
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.4.1
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.67.0
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
httpx-sse==0.4.0
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.0.200
langchain-core==0.2.41
langchain-google-genai==1.0.1
langchainplus-sdk==0.0.20
langgraph==0.2.0
langgraph-checkpoint==1.0.12
langgraph-sdk==0.1.34
langsmith==0.1.137
marshmallow==3.23.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
numexpr==2.10.1
numpy==1.26.4
openapi-schema-pydantic==1.2.4
orjson==3.10.10
packaging==23.2
passlib==1.7.4
pillow==11.0.0
propcache==0.2.0
proto-plus==1.25.0
protobuf==4.25.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==1.10.11
pydantic_core==2.23.4
pyparsing==3.2.0
pypdf==5.0.1
PyPDF2==3.0.1
python-dotenv==1.0.1
python-multipart==0.0.16
PyYAML==6.0.2
reportlab==4.2.5
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.27.0
tenacity==8.5.0
tqdm==4.66.5
typing-inspect==0.9.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.22.0
yarl==1.16.0